系统用户接口

+++ hmx-core/src/main/java/com/hmx/user/dao/UserModelMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Select;
 
 public interface UserModelMapper {
-    long countByExample(UserModelExample example);
+    int countByExample(UserModelExample example);
 
     int deleteByExample(UserModelExample example);
